Added doc strings, changed global defaults in datamodel to be more intelligent and did some styling
